fix: not expose serializeAction, remove serializeActions
