+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from './errors'
 export * from './interfaces'
 export * from './Registrar'
-export * from './SpecAction'
 export * from './SpecExpectation'
